+++ lib/l10n/app_localizations_de.dart
@@ -911,6 +911,11 @@ class AppLocalizationsDe extends AppLocalizations {
   }
 
   @override
+  String csvImportTracks(int count) {
+    return '$count tracks from CSV';
+  }
+
+  @override
   String snackbarAddedToQueue(String trackName) {
     return 'Added \"$trackName\" to queue';
   }
@@ -1583,6 +1588,15 @@ class AppLocalizationsDe extends AppLocalizations {
 
   @override
   String get trackReleaseDate => 'Release date';
+
+  @override
+  String get trackGenre => 'Genre';
+
+  @override
+  String get trackLabel => 'Label';
+
+  @override
+  String get trackCopyright => 'Copyright';
 
   @override
   String get trackDownloaded => 'Downloaded';